refactor: 불필요한 클래스 정리
- 사용자 예외(WrongFindDataException) 삭제 -> 사용안함
- CustomerConverter, VoucherConverter -> Converter 통합
