--- src/main/java/com/programmers/kwonjoosung/springbootbasicjoosung/exception/WrongFindDataException.java
@@ -1,14 +1,0 @@
-package com.programmers.kwonjoosung.springbootbasicjoosung.exception;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-public class WrongFindDataException extends RuntimeException {
-
-    private static final Logger logger = LoggerFactory.getLogger(WrongFindDataException.class);
-
-    public WrongFindDataException(String message) {
-        super(message);
-        logger.error(message);
-    }
-}
